requirements.txt
image-classifiers
keras
numpy==1.17.1
opencv_python==3.4.3.18
sklearn
tensorflow
typing==3.6.6
tensorboard
tensorboard_plugin_profile
Pillow
sklearn
keras_flops
matplotlib